Changed to authOptions model for faster responses
